services/tasks_api/pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["Riad Elimemmedov <riad.elimemmedov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
httpx = "^0.26.0"
mangum = "^0.17.0"
boto3 = "^1.34.11"
python-decouple = "^3.8"
pyjwt = "^2.8.0"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"
bandit = "^1.7.6"
moto = "^4.2.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
